requirements.txt
Django==1.10.6
django-filter==1.0.2
django-webpack-loader==0.4.1
djangorestframework==3.6.2
Markdown==2.6.8
psycopg2==2.7.1
lxml==3.7.3
git+git://github.com/mrhadenough/django-sabridge@python3_compatibility
requests==2.13.0
pycrypto==2.6.1
